[Change] operation_thread to controller_app (implementation code and test cases)
